requirements.txt
# DataCollector 依赖

# 核心依赖
requests>=2.28.0
PyYAML>=6.0

# Web 框架
flask>=2.3.0
flask-cors>=4.0.0

# 数据处理
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 数据库
# sqlite3 (内置)

# 文件处理
python-docx>=0.8.11
openpyxl>=3.1.0
PyMuPDF>=1.22.0  # PDF 处理

# 可选依赖（按需安装）
# pymysql>=1.1.0  # MySQL 支持
# psycopg2-binary>=2.9.0  # PostgreSQL 支持

# 开发工具
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
